restructure apps and models for biodiversity, taxonomy, and reports
- Renamed `apps/trees` to `apps/biodiversity` and updated models to
  reflect biodiversity-specific data.
- Renamed  `apps/observations` to `apps/reports` and updated models
  for measurements and observations.
- Added a new `apps/taxonomy` app to manage taxonomic data,
  including families, genera, and species.
- Updated migrations to reflect the new structure and relationships.
- Removed deprecated models and migrations from `apps/observations`
  and `apps/trees`.
