+++ backend/apps/taxonomy/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ObservationsConfig(AppConfig):
+class TaxonomyConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.observations"
+    name = "apps.taxonomy"
